Feature/task 11

+++ src/main/java/jp/co/seattle/library/controller/DeleteBookController.java
@@ -13,34 +13,47 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import jp.co.seattle.library.service.BooksService;
+import jp.co.seattle.library.service.RentService;
 
 /**
  * 削除コントローラー
  */
 @Controller //APIの入り口
 public class DeleteBookController {
-    final static Logger logger = LoggerFactory.getLogger(DeleteBookController.class);
-    @Autowired
-    private BooksService booksService;
-    
-    /**
-     * 対象書籍を削除する
-     *
-     * @param locale ロケール情報
-     * @param bookId 書籍ID
-     * @param model モデル情報
-     * @return 遷移先画面名
-     */
-    @Transactional
-    @RequestMapping(value = "/deleteBook", method = RequestMethod.POST)
-    public String deleteBook(
-            Locale locale,
-            @RequestParam("bookId") Integer bookId,
-            Model model) {
-        logger.info("Welcome delete! The client locale is {}.", locale);
-        booksService.deleteBook(bookId);
-        model.addAttribute("bookList", booksService.getBookList());        
-        return "home";
-    }
+	final static Logger logger = LoggerFactory.getLogger(DeleteBookController.class);
+	@Autowired
+	private BooksService booksService;
+
+	@Autowired
+	private RentService rentService;
+
+	/**
+	 * 対象書籍を削除する
+	 *
+	 * @param locale ロケール情報
+	 * @param bookId 書籍ID
+	 * @param model モデル情報
+	 * @return 遷移先画面名
+	 */
+	@Transactional
+	@RequestMapping(value = "/deleteBook", method = RequestMethod.POST)
+	public String deleteBook(
+			Locale locale,
+			@RequestParam("bookId") int bookId,
+			Model model) {
+		logger.info("Welcome delete! The client locale is {}.", locale);
+
+		if(rentService.countValues(bookId) == 0) {
+			booksService.deleteBook(bookId);
+			model.addAttribute("bookList", booksService.getBookList());        
+		}else {
+			model.addAttribute("rentingMessage","貸出中のため削除できません。");
+			model.addAttribute("bookDetailsInfo", booksService.getBookInfo(bookId));
+			return "details";
+		}
+
+
+		return "home";
+	}
 
 }
